fix string to properly listen pin updates and move notifiers
