+++ app/lib/features/pins/providers/pins_provider.dart
@@ -1,29 +1,19 @@
 import 'dart:core';
 
 import 'package:acter/common/providers/space_providers.dart';
-import 'package:acter/features/home/providers/client_providers.dart';
+import 'package:acter/features/pins/providers/notifiers/pins_notifiers.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-class AsyncSpacePinsNotifier
-    extends AutoDisposeFamilyAsyncNotifier<List<ActerPin>, Space> {
-  late Stream<void> _listener;
+final pinsProvider =
+    AsyncNotifierProvider.autoDispose<AsyncPinsNotifier, List<ActerPin>>(
+  () => AsyncPinsNotifier(),
+);
 
-  Future<List<ActerPin>> _getPins() async {
-    return (await arg.pins()).toList(); // this might throw internally
-  }
-
-  @override
-  Future<List<ActerPin>> build(Space arg) async {
-    final client = ref.watch(alwaysClientProvider);
-    final spaceId = arg.getRoomId();
-    _listener = client.subscribeStream('$spaceId::PINS'); // stay up to date
-    _listener.forEach((e) async {
-      state = await AsyncValue.guard(() => _getPins());
-    });
-    return _getPins();
-  }
-}
+final pinProvider = AsyncNotifierProvider.autoDispose
+    .family<AsyncPinNotifier, ActerPin, String>(
+  () => AsyncPinNotifier(),
+);
 
 final spacePinsProvider = AsyncNotifierProvider.autoDispose
     .family<AsyncSpacePinsNotifier, List<ActerPin>, Space>(
@@ -42,57 +32,3 @@ final pinnedLinksProvider = FutureProvider.autoDispose
   final pins = await ref.watch(spacePinsProvider(space).future);
   return pins.where((element) => element.isLink()).toList();
 });
-
-class AsyncPinsNotifier extends AutoDisposeAsyncNotifier<List<ActerPin>> {
-  late Stream<void> _listener;
-
-  Future<List<ActerPin>> _getPins() async {
-    final client = ref.watch(alwaysClientProvider);
-    return (await client.pins()).toList(); // this might throw internally
-  }
-
-  @override
-  Future<List<ActerPin>> build() async {
-    final client = ref.watch(alwaysClientProvider);
-    _listener = client.subscribeStream('PINS'); // stay up to date
-    _listener.forEach((e) async {
-      state = await AsyncValue.guard(() => _getPins());
-    });
-    return _getPins();
-  }
-}
-
-final pinsProvider =
-    AsyncNotifierProvider.autoDispose<AsyncPinsNotifier, List<ActerPin>>(
-  () => AsyncPinsNotifier(),
-);
-
-class AsyncPinNotifier
-    extends AutoDisposeFamilyAsyncNotifier<ActerPin, String> {
-  late Stream<void> _listener;
-
-  Future<ActerPin> _getPin() async {
-    final client = ref.watch(alwaysClientProvider);
-    try {
-      return await client.pin(arg);
-    } catch (e) {
-      return await client.waitForPin(arg, null);
-    }
-    // this might throw internally
-  }
-
-  @override
-  Future<ActerPin> build(String arg) async {
-    final client = ref.watch(alwaysClientProvider);
-    _listener = client.subscribeStream(arg); // stay up to date
-    _listener.forEach((e) async {
-      state = await AsyncValue.guard(() => _getPin());
-    });
-    return _getPin();
-  }
-}
-
-final pinProvider = AsyncNotifierProvider.autoDispose
-    .family<AsyncPinNotifier, ActerPin, String>(
-  () => AsyncPinNotifier(),
-);
